requirements.txt
tokenizers==0.15.1
transformers==4.37.2
accelerate==0.21.0
torchmetrics==1.2.0
seaborn==0.13.2
pydantic==2.4.2
textattack==0.3.10
spacy==3.7.2
einops
sentence_transformers==3.3.1
datasets==3.2.0
protobuf==3.20.1